refactor: move tool groups
