+++ src/McpResponse.ts
@@ -26,19 +26,13 @@ import type {McpContext} from './McpContext.js';
 import type {McpPage} from './McpPage.js';
 import {UncaughtError} from './PageCollector.js';
 import {TextSnapshot} from './TextSnapshot.js';
-import {
-  DevTools,
-  getToonEncode,
-  getGcfEncode,
-  type Protocol,
-} from './third_party/index.js';
+import {DevTools, getToonEncode, getGcfEncode} from './third_party/index.js';
 import type {
   ConsoleMessage,
   ImageContent,
   Page,
   ResourceType,
   TextContent,
-  JSONSchema7Definition,
   Extension,
   HTTPRequest,
 } from './third_party/index.js';
@@ -65,155 +59,6 @@ interface TraceInsightData {
   trace: TraceResult;
   insightSetId: string;
   insightName: InsightName;
-}
-
-export function replaceHtmlElementsWithUids(schema: JSONSchema7Definition) {
-  if (typeof schema === 'boolean') {
-    return;
-  }
-
-  let isHtmlElement = false;
-  for (const [key, value] of Object.entries(schema)) {
-    if (key === 'x-mcp-type' && value === 'HTMLElement') {
-      isHtmlElement = true;
-      break;
-    }
-  }
-
-  if (isHtmlElement) {
-    schema.properties = {uid: {type: 'string'}};
-    schema.required = ['uid'];
-  }
-
-  if (schema.properties) {
-    for (const key of Object.keys(schema.properties)) {
-      replaceHtmlElementsWithUids(schema.properties[key]);
-    }
-  }
-
-  if (schema.items) {
-    if (Array.isArray(schema.items)) {
-      for (const item of schema.items) {
-        replaceHtmlElementsWithUids(item);
-      }
-    } else {
-      replaceHtmlElementsWithUids(schema.items);
-    }
-  }
-
-  if (schema.anyOf) {
-    for (const s of schema.anyOf) {
-      replaceHtmlElementsWithUids(s);
-    }
-  }
-  if (schema.allOf) {
-    for (const s of schema.allOf) {
-      replaceHtmlElementsWithUids(s);
-    }
-  }
-  if (schema.oneOf) {
-    for (const s of schema.oneOf) {
-      replaceHtmlElementsWithUids(s);
-    }
-  }
-}
-
-async function getToolGroups(page: McpPage): Promise<ToolGroups> {
-  // Check if there is a `devtoolstooldiscovery` event listener
-  const windowHandle = await page.pptrPage.evaluateHandle(() => window);
-  // @ts-expect-error internal API
-  const client = page.pptrPage._client();
-  const {listeners}: {listeners: Protocol.DOMDebugger.EventListener[]} =
-    await client.send('DOMDebugger.getEventListeners', {
-      objectId: windowHandle.remoteObject().objectId,
-    });
-  if (listeners.find(l => l.type === 'devtoolstooldiscovery') === undefined) {
-    return [];
-  }
-
-  const toolGroups = await page.pptrPage.evaluate(() => {
-    if (window.__dtmcp) {
-      window.__dtmcp.toolGroups = [];
-    }
-    return new Promise<ToolGroups>(resolve => {
-      const event = new CustomEvent('devtoolstooldiscovery');
-      const groups: ToolGroups = [];
-      // @ts-expect-error Adding custom property
-      event.respondWith = toolGroup => {
-        if (!window.__dtmcp) {
-          window.__dtmcp = {};
-        }
-        if (!window.__dtmcp.toolGroups) {
-          window.__dtmcp.toolGroups = [];
-        }
-
-        if (
-          typeof toolGroup.name !== 'string' ||
-          (toolGroup.description &&
-            typeof toolGroup.description !== 'string') ||
-          !Array.isArray(toolGroup.tools)
-        ) {
-          console.error('Invalid toolGroup:', toolGroup);
-          return;
-        }
-        for (const tool of toolGroup.tools) {
-          if (
-            typeof tool.name !== 'string' ||
-            typeof tool.description !== 'string' ||
-            typeof tool.inputSchema !== 'object' ||
-            typeof tool.execute !== 'function'
-          ) {
-            console.error('Invalid tool:', tool);
-            return;
-          }
-        }
-
-        window.__dtmcp.toolGroups.push(toolGroup);
-
-        // When receiving a toolGroup for the first time, expose a simple execution helper
-        if (!window.__dtmcp.executeTool) {
-          window.__dtmcp.executeTool = async (toolName, args) => {
-            if (
-              !window.__dtmcp?.toolGroups ||
-              window.__dtmcp.toolGroups.length === 0
-            ) {
-              throw new Error('No tools found on the page');
-            }
-            for (const group of window.__dtmcp.toolGroups) {
-              const tool = group.tools?.find(t => t.name === toolName);
-              if (tool) {
-                return await tool.execute(args);
-              }
-            }
-            throw new Error(`Tool ${toolName} not found`);
-          };
-        }
-
-        groups.push(toolGroup);
-      };
-      window.dispatchEvent(event);
-      // If at least one toolGroup was added synchronously, resolve with the array.
-      // Otherwise, use setTimeout to allow for any microtask/asynchronous respondWith calls, or resolve with an empty array.
-      if (groups.length > 0) {
-        resolve(groups);
-      } else {
-        setTimeout(() => {
-          if (groups.length > 0) {
-            resolve(groups);
-          } else {
-            resolve([]);
-          }
-        }, 0);
-      }
-    });
-  });
-
-  for (const group of toolGroups) {
-    for (const tool of group.tools ?? []) {
-      replaceHtmlElementsWithUids(tool.inputSchema);
-    }
-  }
-  return toolGroups;
 }
 
 export class McpResponse implements Response {
@@ -685,7 +530,7 @@ export class McpResponse implements Response {
       this.#listThirdPartyDeveloperTools &&
       this.#page
     ) {
-      thirdPartyDeveloperTools = await getToolGroups(this.#page);
+      thirdPartyDeveloperTools = await this.#page.getToolGroups();
       if (thirdPartyDeveloperTools) {
         this.#page.thirdPartyDeveloperTools = thirdPartyDeveloperTools;
       }
